lectin-glycan-interaction-ml/requirements.txt
torch
fair-esm
numpy
scikit-learn
rdkit-pypi
# Optional for GCN encoder
torch-geometric
